requirements.txt
flask
Pillow
werkzeug
joblib
sklearn
numpy
gunicorn